Imported Upstream version 1.6.0

+++ bindings/ruby/main.c
@@ -5,7 +5,7 @@
 #include <unistd.h>
 #include <ruby.h>
 #include <math.h>
-#include "../../src/rrd_tool.h"
+#include "rrd_tool.h"
 
 typedef struct string_arr_t {
     int       len;
@@ -51,13 +51,13 @@ string_arr string_arr_new(
             a.strings[i + 1] = strdup(StringValuePtr(v));
             break;
         case T_FIXNUM:
-            snprintf(buf, 63, "%ld", (long)FIX2INT(v));
+            snprintf(buf, 63, "%d", FIX2INT(v));
             a.strings[i + 1] = strdup(buf);
             break;
         default:
             rb_raise(rb_eTypeError,
-                     "invalid argument - %s, expected T_STRING or T_FIXNUM on index %ld",
-                     rb_class2name(CLASS_OF(v)), (long)i);
+                     "invalid argument - %s, expected T_STRING or T_FIXNUM on index %d",
+                     rb_class2name(CLASS_OF(v)), i);
             break;
         }
     }
